Fix xml2po.pc to not Require libxml.pc
.pc files are really about development packages; for people building
distributions or operating systems, which have split runtime and
devel, we really don't want to pull in libxml2 headers when all that
xml2po needs is the runtime library.

Fixing this is important because I have changed jhbuild to call
pkg-config --modversion on installed pkg-config packages, but it blows
up if there's a missing dependency.

https://bugzilla.gnome.org/show_bug.cgi?id=657339