build: Add --disable-documentation option
This may sound weird for gnome-doc-utils, but basically I need a
way to bootstrap an operating system without documentation initially,
because it's the source of big dependency cycles.

This configure option is enough to get the m4 macros installed, and
then move from there to building GDM without documentation too.