Adding back RNG files for now
Per an IRC discussion with Michael Biebl. The build tools have
needed the RNG files. I removed the RNG files and didn't remove
that requirement. That was a big screw-up. Then I removed the
requirement. That's cool, but there's a whole generation of
packages that still need the RNG without the RNG being there.

So I'm adding the RNG back, but without any rules that call
them. I'll hold it this was for one release cycle, as a way
to transition over.