figures: link to absolute path
While installing figures, we assume to find them in C-language if
they are not explicit part of $lang subfolders.
linking to ../../C/$fig works out well as long as $fig is not living
in any subfolder though. With subfolders, two parent levels is no
longer sufficient, resulting in dangling symlinks

As a solution, we change the link target path to the absolute
destination where we install the C-language figures in. After all, we do
know where we put our files.

Fixes bug 682776.